create workManager
