+++ app/src/main/java/com/example/cryptoapp/domain/LoadDataUseCase.kt
@@ -4,5 +4,5 @@ class LoadDataUseCase(
     private val repository: CoinRepository
 ) {
 
-    suspend operator fun invoke() = repository.loadData()
+    operator fun invoke() = repository.loadData()
 }
